test(gcp): satisfy full validation gates
