+++ apps/api/src/services/gcp-setup.ts
@@ -1,10 +1,10 @@
 import type { GcpOidcCredential } from '@simple-agent-manager/shared';
 import {
   DEFAULT_GCP_API_TIMEOUT_MS,
-  GCP_CREDENTIAL_VERSION,
   DEFAULT_GCP_SERVICE_ACCOUNT_ID,
   DEFAULT_GCP_WIF_POOL_ID,
   DEFAULT_GCP_WIF_PROVIDER_ID,
+  GCP_CREDENTIAL_VERSION,
 } from '@simple-agent-manager/shared';
 import * as v from 'valibot';
