+++ packages/shared/src/constants/index.ts
@@ -161,12 +161,12 @@ export {
 
 // GCP Deployment
 export {
+  DEFAULT_GCP_COMPUTE_API_BASE_URL,
   DEFAULT_GCP_DEPLOY_IDENTITY_TOKEN_EXPIRY_SECONDS,
   DEFAULT_GCP_DEPLOY_OAUTH_STATE_TTL_SECONDS,
   DEFAULT_GCP_DEPLOY_OAUTH_TOKEN_HANDLE_TTL_SECONDS,
   DEFAULT_GCP_DEPLOY_SERVICE_ACCOUNT_ID,
   DEFAULT_GCP_DEPLOY_WIF_POOL_ID,
-  DEFAULT_GCP_COMPUTE_API_BASE_URL,
   DEFAULT_GCP_DEPLOY_WIF_PROVIDER_ID,
   DEFAULT_GCP_IAM_CREDENTIALS_BASE_URL,
   DEFAULT_GCP_SERVICE_ACCOUNT_ASSERTION_LIFETIME_SECONDS,
